remove charStart and charEnd from Findings; further refactoring

+++ coderadar-core/src/main/java/io/reflectoring/coderadar/analyzer/domain/MetricValue.java
@@ -10,7 +10,7 @@ import lombok.Data;
 @AllArgsConstructor
 public class MetricValue {
   private final String name;
-  private final long value;
+  private final int value;
   private final long commitId;
   private final long fileId;
